requirements.txt
pandas
numpy
matplotlib
seaborn
flask
python.box==6.0.2
ensure
pyYAML
pathlib
-e .
